+++ pkg/providers/property/data_akamai_property_rules_template_test.go
@@ -174,7 +174,7 @@ func TestDataAkamaiPropertyRulesRead(t *testing.T) {
 				Steps: []resource.TestStep{
 					{
 						Config:      loadFixtureString("testdata/TestDSRulesTemplate/template_var_not_found.tf"),
-						ExpectError: regexp.MustCompile(`executing "snippets/sub/another-template.json" at <.options>: map has no entry for key "options"`),
+						ExpectError: regexp.MustCompile(`error replacing variable at "testdata/TestDSRulesTemplate/rules/property-snippets/snippets/sub/another-template.json": could not find variable "options"`),
 					},
 				},
 			})
@@ -578,10 +578,12 @@ func TestConvertToTemplate(t *testing.T) {
 		givenFile    string
 		expectedFile string
 		withError    error
+		varMaps      map[string]interface{}
 	}{
 		"valid conversion": {
 			givenFile:    "template_in.json",
 			expectedFile: "template_out.json",
+			varMaps:      map[string]interface{}{"name": `"templateName"`},
 		},
 		"plain JSON passed": {
 			givenFile:    "plain_json.json",
@@ -594,11 +596,12 @@ func TestConvertToTemplate(t *testing.T) {
 		"multiple vars": {
 			givenFile:    "/snippets/some-template.json",
 			expectedFile: "/property-snippets/some-template-out.json",
+			varMaps:      map[string]interface{}{"enabled": true, "criteriaMustSatisfy": `"all"`},
 		},
 	}
 	for name, test := range tests {
 		t.Run(name, func(t *testing.T) {
-			res, err := convertToTemplate(fmt.Sprintf("%s/%s", templates, test.givenFile))
+			res, err := convertToTemplate(fmt.Sprintf("%s/%s", templates, test.givenFile), test.varMaps)
 			if test.withError != nil {
 				assert.True(t, errors.Is(err, test.withError), "want: %s; got: %s", test.withError, err)
 				return
@@ -617,14 +620,17 @@ func TestStringToTemplate(t *testing.T) {
 		givenFile    string
 		expectedFile string
 		withError    error
+		varMaps      map[string]interface{}
 	}{
 		"valid conversion": {
 			givenFile:    "template_in.json",
 			expectedFile: "template_out.json",
+			varMaps:      map[string]interface{}{"name": `"templateName"`},
 		},
 		"multiple includes in array": {
 			givenFile:    "template_in_with_array.json",
 			expectedFile: "template_out_with_array.json",
+			varMaps:      map[string]interface{}{"name": `"templateName"`},
 		},
 		"plain JSON passed": {
 			givenFile:    "plain_json.json",
@@ -634,7 +640,7 @@ func TestStringToTemplate(t *testing.T) {
 	for name, test := range tests {
 		t.Run(name, func(t *testing.T) {
 			givenString := loadFixtureString(fmt.Sprintf("%s/%s", templates, test.givenFile))
-			res, err := stringToTemplate(givenString)
+			res, err := stringToTemplate(givenString, test.varMaps, "main")
 			fmt.Println(res)
 			if test.withError != nil {
 				assert.True(t, errors.Is(err, test.withError), "want: %s; got: %s", test.withError, err)
@@ -643,6 +649,88 @@ func TestStringToTemplate(t *testing.T) {
 			require.NoError(t, err)
 			expected := loadFixtureString(fmt.Sprintf("%s/%s", templatesOut, test.expectedFile))
 			assert.Equal(t, expected, res)
+		})
+	}
+}
+
+func TestVariablesNesting(t *testing.T) {
+	tests := map[string]struct {
+		configPath   string
+		expectedPath string
+	}{
+		"variable is build of other variables": {
+			configPath:   "testdata/TestDSRulesTemplate/template_variables_build.tf",
+			expectedPath: "testdata/TestDSRulesTemplate/output/template_simple.json",
+		},
+		"simple include with variables": {
+			configPath:   "testdata/TestDSRulesTemplate/template_include_with_variables.tf",
+			expectedPath: "testdata/TestDSRulesTemplate/output/temple_with_includes.json",
+		},
+		"multiple includes on the same level": {
+			configPath:   "testdata/TestDSRulesTemplate/template_nested_includes.tf",
+			expectedPath: "testdata/TestDSRulesTemplate/output/template_with_nested_includes.json",
+		},
+		"json list": {
+			configPath:   "testdata/TestDSRulesTemplate/template_with_list.tf",
+			expectedPath: "testdata/TestDSRulesTemplate/output/template_with_list.json",
+		},
+		"variable in include": {
+			configPath:   "testdata/TestDSRulesTemplate/template_variable_building_in_include.tf",
+			expectedPath: "testdata/TestDSRulesTemplate/output/template_include_with_variables.json",
+		},
+		"include child has child": {
+			configPath:   "testdata/TestDSRulesTemplate/template_child_with_childs.tf",
+			expectedPath: "testdata/TestDSRulesTemplate/output/template_with_includes_has_includes.json",
+		},
+	}
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			client := papi.Mock{}
+			useClient(&client, nil, func() {
+				resource.UnitTest(t, resource.TestCase{
+					ProviderFactories: testAccProviders,
+					Steps: []resource.TestStep{
+						{
+							Config: loadFixtureString(test.configPath),
+							Check: resource.ComposeAggregateTestCheckFunc(
+								resource.TestCheckResourceAttr("data.akamai_property_rules_template.test", "json", loadFixtureString(test.expectedPath)),
+							),
+						},
+					},
+				})
+			})
+		})
+	}
+}
+
+func TestVariablesAndIncludesNestingCyclicDependency(t *testing.T) {
+	tests := map[string]struct {
+		configPath string
+		withError  string
+	}{
+		"simple variable cyclic dependency": {
+			configPath: "testdata/TestDSRulesTemplate/template_simple_cyclic_dependency.tf",
+			withError:  "hit cyclic dependency ending at .+",
+		},
+		"tricky variable cyclic dependency": {
+			configPath: "testdata/TestDSRulesTemplate/template_tricky_cyclic_dependency.tf",
+			withError:  "hit cyclic dependency ending at .+",
+		},
+	}
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			client := papi.Mock{}
+			useClient(&client, nil, func() {
+				resource.UnitTest(t, resource.TestCase{
+					ProviderFactories: testAccProviders,
+					Steps: []resource.TestStep{
+						{
+							Config:      loadFixtureString(test.configPath),
+							ExpectError: regexp.MustCompile(test.withError),
+						},
+					},
+				})
+			})
 		})
 	}
 }