+++ pkg/providers/property/testdata/TestDSRulesTemplate/template_child_with_childs.tf
@@ -1,0 +1,22 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+data "akamai_property_rules_template" "test" {
+  template_file = "testdata/TestDSRulesTemplate/child-with-child-rules/main.json"
+  variables {
+    name  = "variableName"
+    value = "simpleName"
+    type  = "string"
+  }
+  variables {
+    name  = "include1Name"
+    value = "include1.json"
+    type  = "string"
+  }
+  variables {
+    name  = "include2Name"
+    value = "include2.json"
+    type  = "string"
+  }
+}