+++ pkg/providers/property/testdata/TestDSRulesTemplate/template_include_with_variables.tf
@@ -1,0 +1,17 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+data "akamai_property_rules_template" "test" {
+  template_file = "testdata/TestDSRulesTemplate/variables-with-includes-rules/main.json"
+  variables {
+    name  = "variableName"
+    value = "simpleName"
+    type  = "string"
+  }
+  variables {
+    name  = "includeName"
+    value = "include1.json"
+    type  = "string"
+  }
+}